add theme styling + enhance UX

+++ src/App.js
@@ -15,38 +15,48 @@ export default {
     }
   },
   methods: {
-    addNewMessage:function() {
-      addDoc(collection(db,'messages'),{
-        text:this.$refs.newmessage.value,
-        date:Date.now(),
-      });  
-      this.$refs.newmessage.value = ""
+    addNewMessage: function () {
+      const hasContent = this.$refs.newmessage.value !== '';
+      hasContent && addDoc(collection(db, 'messages'), {
+        text: this.$refs.newmessage.value,
+        date: Date.now(),
+      });
+      this.$refs.newmessage.value = '';
     },
-    startEditing() {
-      this.isEditing = true;
+
+    startEditing: function (message) {
+      this.messages.forEach((msg) => {
+        msg.isEditing = msg.id === message.id;
+      });
     },
-    updateMessage:function(message) {
-      this.isEditing = false;
-      setDoc(doc(db, 'messages', message.id),{
-        text:message.text,
-        date:message.date
-      })
+
+    updateMessage: function (message) {
+      message.isEditing = false;
+      setDoc(doc(db, 'messages', message.id), {
+        text: message.text,
+        date: message.date,
+      });
     },
-    deleteMessage:function(id) {
-      deleteDoc(doc(db, 'messages', id))
+
+    deleteMessage: function (id) {
+      const isConfirmed = window.confirm("Are you sure you want to delete this message?");
+      isConfirmed && deleteDoc(doc(db, 'messages', id));
     },
   },
+
   mounted() {
-    const latestQuery = query(collection(db, "messages"),orderBy('date'));
-    const livemessages = onSnapshot(latestQuery,(snapshot)=>{
+    const latestQuery = query(collection(db, 'messages'), orderBy('date'));
+    const livemessages = onSnapshot(latestQuery, (snapshot) => {
       this.messages = snapshot.docs.map((doc) => {
         return {
-          id:doc.id,
-          text:doc.data().text,
-          date:doc.data().date
-        }
+          id: doc.id,
+          text: doc.data().text,
+          date: doc.data().date,
+          isEditing: false,
+        };
       });
     });
-    onUnmounted(livemessages)
-  }
+ 
+    onUnmounted(livemessages);
+  },
 }
